رفکتور و کامنت گذاری

+++ src/Http/Controllers/Resource/ResourceSoftDeletingController.php
@@ -13,15 +13,20 @@
             $col = null;
             foreach(request()->resourceId as $id){
 
+                /**
+                 * نظر مورد رکورد کردن پیدا
+                 */
+                $find = $request->getModelAndWhereTrashed()->where([$request->getModelKeyName() => $id])->first();
 
-                $model_query = $this->check_isForceDeleting($request->newModel());
+                /**
+                 * رکورد حذف و دستررسی سطع بررسی
+                 */
+                if(!is_null($find) && $request->Resource()->authorizedToDelete($find))
+                    $find->delete();
 
-                $find = $model_query->where([$request->newModel()->getKeyName() => $id])->first();
-
-                if(!is_null($find))
-                    if($request->Resource()->authorizedToDelete($find))
-                        $find->delete();
-
+                /**
+                 * ها Action دوباره گرفتن
+                 */
                 $col = \Zoroaster::ResourceActions($request , $find ,
                     $request->newModel() , 'Index' , $request->ResourceFields(function($field){
                         if($field !== null && $field->showOnIndex == true)
@@ -42,21 +47,9 @@
 
 
             if(request()->has('redirect'))
-                redirect(request()->redirect)->with([
-                    'success' => 'منبع مورد نظر حذف شد' ,
-                ])->send();
+                redirect(request()->redirect)->with(['success' => 'منبع مورد نظر حذف شد' ])->send();
 
             return response($cols);
-        }
-
-        private function check_isForceDeleting($model)
-        {
-
-            if(method_exists($model , 'isForceDeleting'))
-                return $model->withTrashed();
-            else
-                return $model;
-
         }
 
     }
